requirements.txt
asgiref==3.7.2
asttokens==2.2.1
atomicwrites==1.4.1
attrs==23.1.0
backcall==0.2.0
black==23.7.0
click==8.1.6
colorama==0.4.6
decorator==5.1.1
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
executing==1.2.0
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.7.0
jedi==0.18.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
matplotlib-inline==0.1.6
more-itertools==10.1.0
mypy-extensions==1.0.0
packaging==22.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
pluggy==0.13.1
prompt-toolkit==3.0.36
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.10.0
Pygments==2.13.0
PyJWT==2.6.0
pytest==7.4.0
pytest-django==4.5.2
pytest-testdox==3.0.1
python-dotenv==1.0.0
pytz==2022.6
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
toml==0.10.2
traitlets==5.7.0
tzdata==2023.3
uritemplate==4.1.1
wcwidth==0.2.5
